update, sembra funzionare

+++ src/Op_server.java
@@ -1,35 +1,34 @@
-import javax.naming.PartialResultException;
 import java.io.*;
 import java.net.Socket;
-import java.util.Collection;
 import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.Objects;
 
 @SuppressWarnings("InfiniteLoopStatement")
-public class op_server implements Runnable{
+public class Op_server implements Runnable{
 
     private Socket socket;
     private String op;
     private String username = null;
-    private BufferedWriter writer;
-    private BufferedReader reader;
     private String access_denided ="Accesso negato per questa operazione";
     private String save_failed = "impossibile salvare in maniera persistete i dati";
+    private BufferedWriter writer;
 
-    public op_server(Socket socket){
+
+
+    public Op_server(Socket socket){
         this.socket = socket;
     }
 
 
     @Override
     public void run() {
-
         while(true){
 
             try{
-                reader = new BufferedReader(new InputStreamReader(socket.getInputStream()));
+                BufferedReader reader = new BufferedReader(new InputStreamReader(socket.getInputStream()));
                 writer = new BufferedWriter(new OutputStreamWriter(socket.getOutputStream()));
+
 
                 while (!(op = reader.readLine()).isEmpty()){
                     String[] args = op.split(" ");
@@ -75,8 +74,8 @@ public class op_server implements Runnable{
                             cancelproject(args);
                             break;
                     }
-                }
 
+                }
             }catch (IOException e){
                 e.printStackTrace();
             }
@@ -89,7 +88,7 @@ public class op_server implements Runnable{
         String password = args[2];
         String ret = null;
 
-        hash_users obj = singleton_db_utenti.getInstanceUtenti();
+        Hash_users obj = Singleton_db_utenti.getInstanceUtenti();
         if (obj.login(username,password)){
             this.username = username;
             ret = "SUCCESS "+obj.getUsers();
@@ -106,7 +105,7 @@ public class op_server implements Runnable{
         if(this.username == null) ret = access_denided;
         else if(!username.equals(this.username)) ret = access_denided;
         else {
-            hash_users obj = singleton_db_utenti.getInstanceUtenti();
+            Hash_users obj = Singleton_db_utenti.getInstanceUtenti();
             ret = obj.logout(this.username);
             this.username = null;
         }
@@ -118,9 +117,9 @@ public class op_server implements Runnable{
         String projectname = args[1];
         if(this.username == null) ret = access_denided;
         else {
-            hash_project obj = singleton_db_progetti.getInstanceProgetti();
+            Hash_project obj = Singleton_db_progetti.getInstanceProgetti();
             ret = obj.add_project(projectname, username);
-            persistent_data prs = persistent_data.getInstance();
+            Persistent_data prs = Persistent_data.getInstance();
             if (!prs.create_dir(projectname)) ret = save_failed;
             if (!prs.save(prs.getProject_folder() + projectname + "/membri.json", obj.get_project(projectname).getMembri(), LinkedList.class)) ret = save_failed;
         }
@@ -132,13 +131,13 @@ public class op_server implements Runnable{
         String projectname = args[1];
         String username = args[2];
         if(this.username == null) ret = access_denided;
-        else if(!singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
+        else if(!Singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
         else {
-            hash_project obj = singleton_db_progetti.getInstanceProgetti();
+            Hash_project obj = Singleton_db_progetti.getInstanceProgetti();
             ret = obj.add_projectmember(projectname, username);
-            persistent_data prs = persistent_data.getInstance();
+            Persistent_data prs = Persistent_data.getInstance();
             if (!prs.save(prs.getProject_folder() + projectname + "/membri.json", obj.get_project(projectname).getMembri(), LinkedList.class)) ret = save_failed;
-            if (!prs.save(prs.getUser_folder() + username + "utenti.json", singleton_db_utenti.getInstanceUtenti(), hash_users.class)) ret = save_failed;
+            if (!prs.save(prs.getUser_folder() + username + "utenti.json", Singleton_db_utenti.getInstanceUtenti(), Hash_users.class)) ret = save_failed;
         }
         sendanswer(ret);
     }
@@ -147,9 +146,9 @@ public class op_server implements Runnable{
         String ret = null;
         String projectname = args[1];
         if(this.username == null) ret = access_denided;
-        else if(!singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
+        else if(!Singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
         else {
-            hash_project obj = singleton_db_progetti.getInstanceProgetti();
+            Hash_project obj = Singleton_db_progetti.getInstanceProgetti();
             LinkedList<String> membri = obj.show_members(projectname, this.username);
             Iterator<String> iterator = membri.iterator();
             while (iterator.hasNext()) {
@@ -163,9 +162,9 @@ public class op_server implements Runnable{
         String ret = null;
         String projectname = args[1];
         if(this.username == null) ret = access_denided;
-        else if(!singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
+        else if(!Singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
         else {
-            hash_project obj = singleton_db_progetti.getInstanceProgetti();
+            Hash_project obj = Singleton_db_progetti.getInstanceProgetti();
             LinkedList<String> carte = obj.show_cards(projectname, this.username);
             Iterator<String> iterator = carte.iterator();
             while (iterator.hasNext()) {
@@ -180,9 +179,9 @@ public class op_server implements Runnable{
         String projectname = args[1];
         String cardname = args[2];
         if(this.username == null) ret = access_denided;
-        else if(!singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
+        else if(!Singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
         else {
-            hash_project obj = singleton_db_progetti.getInstanceProgetti();
+            Hash_project obj = Singleton_db_progetti.getInstanceProgetti();
             ret = obj.show_card(projectname, cardname);
         }
         sendanswer(ret);
@@ -194,12 +193,12 @@ public class op_server implements Runnable{
         String cardname = args[2];
         String descrizione = args[3];
         if(this.username == null) ret = access_denided;
-        else if(!singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
+        else if(!Singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
         else {
-            hash_project obj = singleton_db_progetti.getInstanceProgetti();
+            Hash_project obj = Singleton_db_progetti.getInstanceProgetti();
             ret = obj.add_card(projectname, cardname, descrizione);
-            persistent_data prs = persistent_data.getInstance();
-            if(!prs.save(prs.getProject_folder()+cardname+".json",singleton_db_progetti.getInstanceProgetti().get_project(projectname).Getcard(cardname),card.class)) ret = save_failed;
+            Persistent_data prs = Persistent_data.getInstance();
+            if(!prs.save(prs.getProject_folder()+projectname+"/"+cardname+".json", Singleton_db_progetti.getInstanceProgetti().get_project(projectname).Getcard(cardname), Card.class)) ret = save_failed;
         }
         sendanswer(ret);
     }
@@ -211,12 +210,12 @@ public class op_server implements Runnable{
         String listapartenza = args[3];
         String listadestinazione = args[4];
         if(this.username == null) ret = access_denided;
-        else if(!singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
+        else if(!Singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
         else {
-            hash_project obj = singleton_db_progetti.getInstanceProgetti();
+            Hash_project obj = Singleton_db_progetti.getInstanceProgetti();
             ret = obj.move_card(projectname, cardname, listapartenza, listadestinazione);
-            persistent_data prs = persistent_data.getInstance();
-            if(!prs.save(prs.getProject_folder()+cardname+".json",singleton_db_progetti.getInstanceProgetti().get_project(projectname).Getcard(cardname),card.class)) ret = save_failed;
+            Persistent_data prs = Persistent_data.getInstance();
+            if(!prs.save(prs.getProject_folder()+projectname+"/"+cardname+".json", Singleton_db_progetti.getInstanceProgetti().get_project(projectname).Getcard(cardname), Card.class)) ret = save_failed;
         }
         sendanswer(ret);
     }
@@ -226,9 +225,9 @@ public class op_server implements Runnable{
         String projectname = args[1];
         String cardname = args[2];
         if(this.username == null) ret = access_denided;
-        else if (!singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
+        else if (!Singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
         else {
-            hash_project obj = singleton_db_progetti.getInstanceProgetti();
+            Hash_project obj = Singleton_db_progetti.getInstanceProgetti();
             ret = obj.get_cardhistory(projectname, cardname);
         }
         sendanswer(ret);
@@ -238,19 +237,21 @@ public class op_server implements Runnable{
         String projectname = args[1];
         String ret = null;
         if(this.username == null) ret = access_denided;
-        else if(!singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
+        else if(!Singleton_db_progetti.getInstanceProgetti().get_project(projectname).containsmember(this.username)) ret = access_denided;
         else {
-            hash_project obj = singleton_db_progetti.getInstanceProgetti();
-            ret = obj.cancell_project(projectname);
-            persistent_data prs = persistent_data.getInstance();
-            try {
-                File directory = new File((prs.getProject_folder() + projectname));
-                for (File file : Objects.requireNonNull(directory.listFiles()))
-                    file.delete();
-                directory.delete();
-            }catch (Exception e){
-                e.printStackTrace();
-                ret = "impossibile eliminare la persistenza del progetto";
+            Hash_project obj = Singleton_db_progetti.getInstanceProgetti();
+            ret = obj.cancell_project(projectname,this.username);
+            if(ret.startsWith("SUCCESS")) {
+                Persistent_data prs = Persistent_data.getInstance();
+                try {
+                    File directory = new File((prs.getProject_folder() + projectname));
+                    for (File file : Objects.requireNonNull(directory.listFiles()))
+                        file.delete();
+                    directory.delete();
+                } catch (Exception e) {
+                    e.printStackTrace();
+                    ret = "impossibile eliminare la persistenza del progetto";
+                }
             }
         }
         sendanswer(ret);
@@ -260,7 +261,7 @@ public class op_server implements Runnable{
         String ret = null;
         if(this.username == null) ret = access_denided;
         else {
-            hash_users obj = singleton_db_utenti.getInstanceUtenti();
+            Hash_users obj = Singleton_db_utenti.getInstanceUtenti();
             ret = obj.get_onlineusers();
         }
         sendanswer(ret);
@@ -270,7 +271,7 @@ public class op_server implements Runnable{
         String ret = null;
         if(this.username == null) ret = access_denided;
         else {
-            hash_users obj = singleton_db_utenti.getInstanceUtenti();
+            Hash_users obj = Singleton_db_utenti.getInstanceUtenti();
             ret = obj.get_listproject(username);
         }
         sendanswer(ret);
@@ -278,7 +279,8 @@ public class op_server implements Runnable{
 
     private void sendanswer(String answer){
         try {
-            writer.write(answer);
+            writer.write(answer + "\n\r\n");
+            writer.flush();
         }catch (IOException e){
             e.printStackTrace();
         }
